adding messaging to proyect

+++ server/src/routers/cart.js
@@ -1,12 +1,18 @@
 const express = require("express");
 const router = express.Router();
 const DAO = require("../../model/cart.model");
+const USERDAO = require("../../model/login.model");
+const messages = require("../utils/messages");
 
 //get all carts
 router.get("/", async (request, resolve) => {
   try {
     const data = await DAO.getAll();
-    resolve.send(data);
+    if (!data) {
+      throw new Error("Carts not found");
+    } else {
+      resolve.send(data);
+    }
   } catch (error) {
     throw new Error(error);
   }
@@ -29,11 +35,15 @@ router.post("/:id", async (request, resolve) => {
 
 //delete cart by id
 router.delete("/:id/products", async (request, resolve) => {
-  const id = request.params.id;
-
   try {
-    const deleted = await DAO.deleteById(id);
-    resolve.send(deleted);
+    const id = request.params.id;
+    const search = await DAO.getCartById(id);
+    if (!search) {
+      throw new Error("Cart not found");
+    } else {
+      const deleted = await DAO.deleteById(id);
+      resolve.send(deleted);
+    }
   } catch (error) {
     throw new Error(error);
   }
@@ -41,14 +51,13 @@ router.delete("/:id/products", async (request, resolve) => {
 
 //get cart by id
 router.get("/:id/products", async (request, resolve) => {
-  const id = request.params.id;
-
   try {
-    const data = await DAO.getIdByCart(id);
-    if (!data) {
-      resolve.send({ error: "cart not found" });
+    const id = request.params.id;
+    const search = await DAO.getCartById(id);
+    if (!search) {
+      throw new Error("Cart not found");
     } else {
-      resolve.send(data);
+      resolve.send(search);
     }
   } catch (error) {
     throw new Error(error);
@@ -57,10 +66,9 @@ router.get("/:id/products", async (request, resolve) => {
 
 //add product to cart
 router.post("/:id/products", async (request, resolve) => {
-  const id = request.params.id;
-  const newData = request.body;
-
   try {
+    const id = request.params.id;
+    const newData = request.body;
     const data = await DAO.editCart(newData, id);
     resolve.send(data);
   } catch (error) {
@@ -78,6 +86,31 @@ router.delete("/:id/products/:id_prod", async (request, resolve) => {
     const deleteProduct = await DAO.deleteProduct(cartID, prodID);
 
     resolve.send({ message: "Product deleted", deleteProduct });
+  } catch (error) {
+    throw new Error(error);
+  }
+});
+
+//order process
+router.post("/:id/shop", async (request, resolve) => {
+  try {
+    const client = request.params.id;
+    const search = await DAO.getByID(client);
+    if (!search) {
+      throw new Error("Cart not found");
+    } else {
+      const { name, email, phone } = await USERDAO.getById(client);
+
+      const subject = "Your order...";
+      const messageOrder = `${name} Your order is in process`;
+      messages.sms(messageOrder, phone);
+      messages.whatsapp(messageOrder, phone);
+      messages.gmail(subject, messageOrder, email);
+      await DAO.deleteByCart(client).then((resolve) => {
+        resolve.json(`Carrito ${client}: Se borró con éxito.`);
+      });
+      resolve.send({ message: "Product in process", search });
+    }
   } catch (error) {
     throw new Error(error);
   }
